Introducing `FieldValue` component using `stringify` helper.
