+++ graylog2-web-interface/src/components/common/message/details/fields/FieldValue.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+import stringify from 'util/stringify';
+
+type Props = { value: number | string | bigint };
+const FieldValue = ({ value }: Props) => <>{stringify(value)}</>;
+export default FieldValue;
